extract WebhookHelper class
