+++ SendgridParquetLogger/Helper/WebhookHelper.cs
@@ -1,0 +1,103 @@
+using System.Buffers;
+using System.IO.Pipelines;
+using System.Net;
+using System.Text;
+using System.Text.Json;
+
+using Microsoft.AspNetCore.Http;
+
+using SendgridParquet.Shared;
+
+using ZLogger;
+
+namespace SendgridParquetLogger.Helper;
+
+public class WebhookHelper(
+    ILogger<WebhookHelper> logger,
+    TimeProvider timeProvider,
+    ParquetService parquetService,
+    RequestValidator requestValidator,
+    S3StorageService s3StorageService
+)
+{
+    private const int MaxBodyBytes = 1 * 1024 * 1024; // 1MB
+
+    public async Task<(HttpStatusCode, ICollection<SendGridEvent>)> ReadSendGridEvents(
+        PipeReader reader,
+        IHeaderDictionary requestHeaders,
+        CancellationToken ct)
+    {
+        ReadResult result = await reader.ReadAtLeastAsync(MaxBodyBytes, ct);
+        ReadOnlySequence<byte> buffer = result.Buffer;
+        if (result.IsCanceled || result.IsCompleted && buffer.Length == 0)
+        {
+            return (HttpStatusCode.BadRequest, Array.Empty<SendGridEvent>());
+        }
+
+        string payload = Encoding.UTF8.GetString(buffer);
+
+        switch (requestValidator.VerifySignature(payload, requestHeaders))
+        {
+            case RequestValidator.RequestValidatorResult.Verified:
+#if DEBUG
+            // DEBUG ビルド時は NotConfigured も許可
+            case RequestValidator.RequestValidatorResult.NotConfigured:
+#endif
+                try
+                {
+                    var events = JsonSerializer.Deserialize<SendGridEvent[]>(payload) ?? [];
+                    return (HttpStatusCode.OK, events);
+                }
+                catch (JsonException)
+                {
+                    // return BadRequest
+                }
+                break;
+        }
+        return (HttpStatusCode.BadRequest, Array.Empty<SendGridEvent>());
+    }
+
+    public async Task<List<HttpStatusCode>> WriteParquetGroupByYmd(
+        IEnumerable<SendGridEvent> events,
+        CancellationToken ct)
+    {
+        DateTimeOffset now = timeProvider.GetUtcNow();
+        var results = new List<HttpStatusCode>(2);
+        foreach (var grp in events
+                     .Select(sendgridEvent => (sendgridEvent, timestamp: JstExtension.JstUnixTimeSeconds(sendgridEvent.Timestamp)))
+                     .GroupBy(pair => new DateOnly(pair.timestamp.Year, pair.timestamp.Month, pair.timestamp.Day), pair => pair.sendgridEvent))
+        {
+            DateOnly targetDay = grp.Key;
+            HttpStatusCode result = await WriteParquet(grp, targetDay, now, ct);
+            results.Add(result);
+        }
+        return results;
+    }
+
+    public async ValueTask<HttpStatusCode> WriteParquet(
+        IEnumerable<SendGridEvent> eventsEnumerable,
+        DateOnly targetDay,
+        DateTimeOffset now,
+        CancellationToken ct)
+    {
+        var events = eventsEnumerable.ToArray();
+        await using var parquetData = await parquetService.ConvertToParquetAsync(events);
+        if (parquetData == null)
+        {
+            logger.ZLogError($"Failed to convert events to Parquet format");
+            return HttpStatusCode.InternalServerError;
+        }
+
+        string fileName = SendGridPathUtility.GetParquetNonCompactionFileName(targetDay, parquetData);
+        bool uploadSuccess = await s3StorageService.PutObjectAsync(parquetData, fileName, now, ct);
+        if (!uploadSuccess)
+        {
+            logger.ZLogError($"Failed to upload Parquet file to S3");
+            return HttpStatusCode.InternalServerError;
+        }
+
+        logger.ZLogInformation($"Successfully processed and stored {targetDay:yyyy/MM/dd} events in {fileName}");
+        return HttpStatusCode.OK;
+    }
+}
+
